+++ src/test/mocks/MockCondition.sol
@@ -9,17 +9,17 @@ contract MockCondition is BaseCondition {
         bool active;
     }
 
-    mapping(address wallet => mapping(uint16 => Condition)) private conditions;
+    mapping(address wallet => mapping(uint32 => Condition)) private conditions;
 
-    function addCondition(uint16 id, Condition calldata condition) external {
+    function addCondition(uint32 id, Condition calldata condition) external {
         conditions[msg.sender][id] = condition;
     }
 
-    function _isConditionActive(address, uint16 _id) internal view override returns (bool) {
+    function _isConditionActive(address, uint32 _id) internal view override returns (bool) {
         return conditions[msg.sender][_id].active;
     }
 
-    function checkCondition(address wallet, uint16 id) public view override returns (uint8) {
+    function checkCondition(address wallet, uint32 id) public view override returns (uint8) {
         return conditions[wallet][id].result ? 1 : 0;
     }
 }
