+++ test/BaseCondition.t.sol
@@ -17,14 +17,14 @@ contract BaseConditionTest is Test {
     /////////  checkCondition /////////
     ///////////////////////////////////
 
-    modifier conditionActivated(uint16 _id) {
+    modifier conditionActivated(uint32 _id) {
         vm.prank(wallet);
         MockCondition.Condition memory _condition = MockCondition.Condition({result: true, active: true});
         condition.addCondition(_id, _condition);
         _;
     }
 
-    function test_checkCondition(uint16 _id) external conditionActivated(_id) {
+    function test_checkCondition(uint32 _id) external conditionActivated(_id) {
         uint8 _result = condition.checkCondition(wallet, _id);
 
         assert(_result == 1);
@@ -34,14 +34,14 @@ contract BaseConditionTest is Test {
     /////////  addStrategyToCondition /////////
     ///////////////////////////////////////////
 
-    function test_addStrategyToCondition_Success(uint16 _id) external conditionActivated(_id) {
+    function test_addStrategyToCondition_Success(uint32 _id) external conditionActivated(_id) {
         uint256 strategyNo = 5;
         for (uint256 i = 0; i < 5; i++) {
             vm.prank(wallet);
-            condition.addStrategyToCondition(_id, uint16(i));
+            condition.addStrategyToCondition(_id, uint32(i));
         }
 
-        uint16[] memory _strategies = condition.strategies(wallet, _id);
+        uint32[] memory _strategies = condition.strategies(wallet, _id);
 
         assertEq(_strategies.length, strategyNo);
     }
